requirements.txt
Flask
Flask-SQLAlchemy
Flask-Session
python-dateutil
python-dotenv
regex
requests
Flask-APScheduler